Use subs instead of global variables for local & binary cache store
Otherwise the `include`-test would fail: this test defines a `hydra.conf`
with an `include` statement. However, the file to be included is created
after the test base is set up. This means that - among others - the
`Hydra::Helper::Nix` module is loaded before the file to include exists.
Since the global variables are directly initialized and try to read from
the config, the test fails, because loading the config at that time
doesn't work. Delaying this into a subrouting solves the issue.

I decided against messing around with the config mechanism here since I
consider this behavior to be correct.

The local store gets (i.e. `Hydra::Store->new()` w/o args) cached in the
Perl bindings already[1], however this isn't the case for the binary
cache store. Because of that, a `state` variable is used to cache the
binary cache store. At runtime, the location of that store doesn't
change anyways.

[1] https://github.com/NixOS/nix/blob/e3ccb7e42a55529d8d50544e7e0cebb5cbc606cf/perl/lib/Nix/Store.xs#L88-L96